Update tests for old replay support
